Add separate file for `RotateContext` command.
